requirements.txt
langchain
chromadb
streamlit
pypdf
python-dotenv
ollama
langsmith
